pyproject.toml
[project]
name = "canonical-agent-architecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autoevals>=0.0.130",
    "braintrust>=0.3.1",
    "openai>=2.1.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]
